fix: trying to fix rerendering on clicking on panel
