+++ admin/src/elements/InputField.jsx
@@ -1,6 +1,7 @@
 import { useI18n } from '@wordpress/react-i18n';
 import { useState, useCallback } from 'react';
 import '../assets/styles/elements/_Inputs.scss';
+import { delay } from '../lib/helpers';
 import Tooltip from './Tooltip';
 
 export default function InputField( { defaultValue, isLoading, autoFocus, placeholder, message, liveUpdate, className, type, readonly, disabled, label, description, labelInline, onChange, onKeyDown, onBlur, onFocus, onKeyUp, children, required, style } ) {
@@ -13,6 +14,12 @@ export default function InputField( { defaultValue, isLoading, autoFocus, placeh
 			onChange( type === 'number' ? event.target.valueAsNumber : event.target.value );
 		}
 	}, [ onChange, type, defaultValue, val ] );
+
+	const handleValLive = ( event ) => {
+		if ( liveUpdate ) {
+			delay( () => handleVal( event ), 800 )();
+		}
+	};
 
 	const valueStatus = () => {
 		if ( val ) {
@@ -48,7 +55,7 @@ export default function InputField( { defaultValue, isLoading, autoFocus, placeh
 							autoFocus={ autoFocus }
 							onChange={ ( event ) => {
 								setVal( event.target.value );
-								handleVal( event );
+								handleValLive( event );
 							} }
 							onBlur={ ( event ) => {
 								handleVal( event ); if ( onBlur ) {
